requirements.txt
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
packaging==24.0
paramiko==3.4.0
pycparser==2.22
PyJWT==2.8.0
PyNaCl==1.5.0
Werkzeug==3.0.3
